requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-google-genai>=1.0.5
langgraph>=0.2.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
plotly>=5.20.0
python-dotenv>=1.0.0
openpyxl>=3.1.0